Reorganise helper functions
Create dedicated module for each function domain.
