+++ src/lib/string.js
@@ -28,3 +28,18 @@ export default {
     return result
   }
 }
+
+export const slugify = (str) => {
+  str = str.replace(/^\s+|\s+$/g, '')
+    .toLowerCase()
+
+  const from = 'àáäâèéëêìíïîòóöôùúüûñç·/_,:'
+  const to = 'aaaaeeeeiiiioooouuuunc------'
+  for (let i = 0, l = from.length; i < l; i++) {
+    str = str.replace(new RegExp(from.charAt(i), 'g'), to.charAt(i))
+  }
+
+  return str.replace(/[^a-z0-9 -]/g, '')
+    .replace(/\s+/g, '_')
+    .replace(/-+/g, '_')
+}
